UIThread Exercise
Completed:
2. Update the code using the Task class from this INCORRECT Threads
tutorial (Links to an external site.) (ignore the for loop - it is
replaced by the takeSomeTime(int) method) . Note which of the Android
Threading Rules is violated.
3. Modify the Android project using this Thread Handler tutorial (Links
to an external site.) which uses a handler to work with thread
correctly. (instead of creating a handler, just use the bar View as the
handler.)